+++ SumupSDK.embeddedframework/SumupSDK.framework/Versions/A/Headers/SumupSDK.h
@@ -27,55 +27,48 @@ typedef void (^SMPCheckoutCompletionBlock)(SMPCheckoutResult * _Nullable result 
 @interface SumupSDK : NSObject
 
 /**
- \abstract Enables the sandbox mode. Default is NO.
- \discussion
- Needs to be called prior to calling +setupWithAPIKey: otherwise does nothing.
- 
- \warning
- You are responsible to ensure that you do not do enable the sandbox 
- mode in production code. Cards will not be charged in sandbox mode.
+ *  Enables the sandbox mode. Default is NO.
+ *
+ *  Needs to be called prior to calling +setupWithAPIKey: otherwise does nothing.
+ * 
+ *  @warning You are responsible to ensure that you do not do enable the sandbox
+ *  mode in production code. Cards will not be charged in sandbox mode.
  */
 + (void)setSandboxModeEnabled:(BOOL)enabled DEPRECATED_MSG_ATTRIBUTE("The sandbox environment is not supported anymore. Please get in touch and we will create a sandboxed test account for you.");
 
 /**
- \return YES if sandbox mode is enabled. Default is NO.
+ *  @return YES if sandbox mode is enabled. Default is NO.
  */
 + (BOOL)isSandboxModeEnabled DEPRECATED_MSG_ATTRIBUTE("see setSandboxModeEnabled:");
 
 /**
- \abstract Sets up the SumupSDK for use in your app.
- \param apiKey Your application's API Key for the SumupSDK.
-  \returns YES if setup was successful. NO otherwise or if SDK has been set up before.
- \discussion
- Needs to be called at some point before starting interaction with the SDK.
- As this might ask for the user's location it should not necessarily be part
- of the app launch. Make sure to only setup once per app lifecycle.
- 
- If the user did not previously grant your app the permission to use her location, 
- calling this method will prompt the user to grant such permission. 
- 
- Please make sure to include the Info.plist localized key 'NSLocationUsageDescription' (the SDK also 
- uses this key on iOS5). The localized string for that key will be displayed in the alertview 
- requesting the user's permission. It is recommended to include a reference to the use of 
- the user's location during the payment process.
- Recommendation: "To provide a secure payment service, we need to know your location. 
- Without location information, you cannot accept payments using this app."
+ *  Sets up the SumupSDK for use in your app.
+ *
+ *  Needs to be called at some point before starting interaction with the SDK.
+ *  As this might ask for the user's location it should not necessarily be part
+ *  of the app launch. Make sure to only setup once per app lifecycle.
+ *
+ *  If the user did not previously grant your app the permission to use her location,
+ *  calling this method will prompt the user to grant such permission.
+ *
+ *  @param apiKey Your application's API Key for the SumupSDK.
+ *  @return YES if setup was successful. NO otherwise or if SDK has been set up before.
  */
 + (BOOL)setupWithAPIKey:(NSString *)apiKey;
 
 
 /**
- \abstract Presents the login modally from the given view controller.
- 
- \param fromViewController The UIViewController instance from which the login should be presented modally.
- \param animated Pass YES to animate the transition.
- \param completionBlock The completion block is called after each login attempt.
- 
- \discussion
- The login is automatically dismissed if login was successful or cancelled by the user.
- If error is nil and success is NO, the user cancelled the login.
- Errors are handled internally and usually do not need any display to the user.
- Does nothing if merchant is already logged in (calls completion block with success=NO, error=nil).
+ *  Presents the login modally from the given view controller.
+ *
+ *  The login is automatically dismissed if login was successful or cancelled by the user.
+ *  If error is nil and success is NO, the user cancelled the login.
+ *  Errors are handled internally and usually do not need any display to the user.
+ *  Does nothing if merchant is already logged in (calls completion block with success=NO, error=nil).
+ *
+ *  @param fromViewController The UIViewController instance from which the login should be presented modally.
+ *  @param animated Pass YES to animate the transition.
+ *  @param block The completion block is called after each login attempt.
+ *
  */
 + (void)presentLoginFromViewController:(UIViewController *)fromViewController
                               animated:(BOOL)animated
@@ -83,8 +76,8 @@ typedef void (^SMPCheckoutCompletionBlock)(SMPCheckoutResult * _Nullable result 
 
 
 /**
- * Logs in a merchant with an access token acquired via https://sumup.com/integration#APIAuth
- * Make sure that no user is logged in already when calling this method.
+ *  Logs in a merchant with an access token acquired via https://sumup.com/integration#APIAuth
+ *  Make sure that no user is logged in already when calling this method.
  *
  *  @param aToken an access token
  *  @param block  a completion block that will run after login has succeeded/failed
@@ -92,22 +85,32 @@ typedef void (^SMPCheckoutCompletionBlock)(SMPCheckoutResult * _Nullable result 
 + (void)loginWithToken:(NSString *)aToken completion:(nullable SumupCompletionBlock)block;
 
 /**
- \returns YES if the merchant is logged in. NO otherwise.
+ *  @return YES if the merchant is logged in. NO otherwise.
  */
 + (BOOL)isLoggedIn;
 
 /// Returns a copy of the currently logged in merchant or nil if no merchant is logged in.
 + (nullable SMPMerchant *)currentMerchant;
 
+
 /**
- \abstract Presents a checkout view with all necessary steps to charge a customer.
- \param request The SMPCheckoutRequest encapsulates all transaction relevant data such as total amount, label, etc.
- \param fromViewController The UIViewController instance from which the checkout should be presented modally.
- \param completionBlock The completion block will be called when the view will be dismissed.
- 
- \discussion
- Does nothing if merchant is not logged in or a checkout is already in progress.
- The completion block is always called.
+ *  Can be called in advance when a checkout is imminent and a user is logged in.
+ *  You should use this method to let the SDK know that the user is most likely starting a
+ *  checkout attempt soon, e.g. when entering an amount or adding products to a shopping cart.
+ *  This allows the SDK to take appropriate measures, like attempting to wake a connected card terminal.
+ */
++ (void)prepareForCheckout;
+
+/**
+ *
+ *  Presents a checkout view with all necessary steps to charge a customer.
+ *
+ *  Does nothing if merchant is not logged in or a checkout is already in progress.
+ *  The completion block is always called.
+ *
+ *  @param request The SMPCheckoutRequest encapsulates all transaction relevant data such as total amount, label, etc.
+ *  @param controller The UIViewController instance from which the checkout should be presented modally.
+ *  @param block The completion block will be called when the view will be dismissed.
  */
 + (void)checkoutWithRequest:(SMPCheckoutRequest *)request
          fromViewController:(UIViewController *)controller
@@ -126,7 +129,7 @@ typedef void (^SMPCheckoutCompletionBlock)(SMPCheckoutResult * _Nullable result 
                  completion:(nullable SMPCheckoutCompletionBlock)block;
 
 /**
- \returns YES if a checkout is progress. NO otherwise.
+ *  @return YES if a checkout is progress. NO otherwise.
  */
 + (BOOL)checkoutInProgress;
 
@@ -148,18 +151,18 @@ typedef void (^SMPCheckoutCompletionBlock)(SMPCheckoutResult * _Nullable result 
                                           completion:(nullable SumupCompletionBlock)block;
 
 /**
- Performs a logout of the current merchant and resets the remembered password.
- \param block The completion block is called once the logout has finished.
+ *  Performs a logout of the current merchant and resets the remembered password.
+ *
+ *  @param block The completion block is called once the logout has finished.
  */
 + (void)logoutWithCompletionBlock:(nullable SumupCompletionBlock)block;
 
 /**
- \abstract If enabled the SDK automatically displays UI notifications (similar to the ringer UI notifications from iOS) to the user when the reader state changes. Default is YES.
- 
- \discussion 
- Notifications are only displayed once the SDK sees fit to do so. If you want to allow UI notifications to appear earlier,
- call this method with YES at any point after +setupWithAPIKey:. Be aware that this will prompt the user to grant permissions
- to your app to use the device's microphone. 
+ *  If enabled the SDK automatically displays UI notifications (similar to the ringer UI notifications from iOS) to the user when the reader state changes. Default is YES.
+ *
+ *  Notifications are only displayed once the SDK sees fit to do so. If you want to allow UI notifications to appear earlier,
+ *  call this method with YES at any point after +setupWithAPIKey:. Be aware that this will prompt the user to grant permissions
+ *  to your app to use the device's microphone.
  */
 + (void)setUINotificationsForReaderStatusEnabled:(BOOL)enabled;
 
@@ -177,13 +180,12 @@ typedef void (^SMPCheckoutCompletionBlock)(SMPCheckoutResult * _Nullable result 
  */
 + (NSString *)bundleVersionShortString;
 
-#pragma mark  - error domain and codes
+#pragma mark - Error Domain and Codes
+
 extern NSString * const SMPSumupSDKErrorDomain;
 
-/*!
- * @enum SMPSumupSDKError
- *
- * @discussion The error codes returned from the SDK
+/**
+ *  The error codes returned from the SDK
  */
 typedef NS_ENUM(NSInteger, SMPSumupSDKError) {
     SMPSumupSDKErrorGeneral             = 0,        // General error
@@ -194,6 +196,19 @@ typedef NS_ENUM(NSInteger, SMPSumupSDKError) {
     SMPSumupSDKErrorCheckoutGeneral     = 50,
     SMPSumupSDKErrorCheckoutInProgress  = 51,       // Another checkout process is currently in progress.
 };
+
+#pragma mark - SDK Integration
+
+/**
+ *  You can use this method to test if you have integrated the SDK correctly.
+ *
+ *  The method will log several tests to the console. If you see any errors, please check the README for setup instructions.
+ *
+ *  @warning While not harmful, this method is not meant to be used in production. Use this in temporary code or in debug configurations only.
+ *
+ *  @return YES if the SDK is set up correctly, NO if any error was found.
+ */
++ (BOOL)testSDKIntegration;
 
 @end