+++ SumupSDKSampleApp/SrcSwift/AppDelegate.swift
@@ -13,6 +13,15 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     var window: UIWindow?
 
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplicationLaunchOptionsKey: Any]?) -> Bool {
+
+        #if DEBUG
+            /*
+             *   Logs integration warnings in non-production code. Do not call this method in
+             *   release builds.
+             */
+            SumupSDK.testIntegration()
+        #endif
+
         /*
          *   This will setup the SumUpSDK.
          *